Fixing crowded Footer
Shortened text to make for a crisp and short footer. Removed copyright
symbol, as it's an opensource project. Removed source on Github link as
we now have a banner for that on Top-Right.

Also removed 'If you don't see anything' message as it should only be
shown when there is nothing to see, not always.